+++ spec/cuprum/collections/basic/commands/assign_one_spec.rb
@@ -3,13 +3,14 @@
 require 'stannum/constraints/types/hash_with_string_keys'
 
 require 'cuprum/collections/basic/commands/assign_one'
-require 'cuprum/collections/basic/rspec/command_contract'
-require 'cuprum/collections/rspec/assign_one_command_contract'
-
-require 'support/examples/basic_command_examples'
+require 'cuprum/collections/rspec/contracts/basic/command_contracts'
+require 'cuprum/collections/rspec/contracts/command_contracts'
 
 RSpec.describe Cuprum::Collections::Basic::Commands::AssignOne do
-  include Spec::Support::Examples::BasicCommandExamples
+  include Cuprum::Collections::RSpec::Contracts::Basic::CommandContracts
+  include Cuprum::Collections::RSpec::Contracts::CommandContracts
+
+  with_contract 'with basic command contexts'
 
   include_context 'with parameters for a basic contract'
 
@@ -21,8 +22,8 @@ RSpec.describe Cuprum::Collections::Basic::Commands::AssignOne do
     )
   end
 
-  let(:initial_attributes)  { {} }
-  let(:entity)              { initial_attributes }
+  let(:initial_attributes) { {} }
+  let(:entity)             { initial_attributes }
   let(:expected_value) do
     SleepingKingStudios::Tools::HashTools
       .instance
@@ -39,8 +40,8 @@ RSpec.describe Cuprum::Collections::Basic::Commands::AssignOne do
     end
   end
 
-  include_contract Cuprum::Collections::Basic::RSpec::COMMAND_CONTRACT
+  include_contract 'should be a basic command'
 
-  include_contract Cuprum::Collections::RSpec::ASSIGN_ONE_COMMAND_CONTRACT,
+  include_contract 'should be an assign one command',
     allow_extra_attributes: true
 end