+++ spec/cuprum/collections/basic/commands/update_one_spec.rb
@@ -2,13 +2,14 @@
 
 require 'cuprum/collections/basic/commands/update_one'
 require 'cuprum/collections/basic/query'
-require 'cuprum/collections/basic/rspec/command_contract'
-require 'cuprum/collections/rspec/update_one_command_contract'
-
-require 'support/examples/basic_command_examples'
+require 'cuprum/collections/rspec/contracts/basic/command_contracts'
+require 'cuprum/collections/rspec/contracts/command_contracts'
 
 RSpec.describe Cuprum::Collections::Basic::Commands::UpdateOne do
-  include Spec::Support::Examples::BasicCommandExamples
+  include Cuprum::Collections::RSpec::Contracts::Basic::CommandContracts
+  include Cuprum::Collections::RSpec::Contracts::CommandContracts
+
+  with_contract 'with basic command contexts'
 
   include_context 'with parameters for a basic contract'
 
@@ -51,9 +52,9 @@ RSpec.describe Cuprum::Collections::Basic::Commands::UpdateOne do
     end
   end
 
-  include_contract Cuprum::Collections::Basic::RSpec::COMMAND_CONTRACT
+  include_contract 'should be a basic command'
 
-  include_contract Cuprum::Collections::RSpec::UPDATE_ONE_COMMAND_CONTRACT
+  include_contract 'should be an update one command'
 
   wrap_context 'with a custom primary key' do
     let(:attributes) do
@@ -62,6 +63,6 @@ RSpec.describe Cuprum::Collections::Basic::Commands::UpdateOne do
         .merge(uuid: '00000000-0000-0000-0000-000000000000')
     end
 
-    include_contract Cuprum::Collections::RSpec::UPDATE_ONE_COMMAND_CONTRACT
+    include_contract 'should be an update one command'
   end
 end